backend/requirements.txt
fastapi==0.115.12
uvicorn==0.30.6
pydantic-settings==2.6.1
httpx==0.27.2
cachetools==5.5.0
firebase-admin==6.5.0
openai==1.52.2
anthropic==0.37.1
email-validator==2.2.0
python-multipart==0.0.22
